Explicit cast f_type to (unsigned int)
This is about the *extremely* annoying __fsword_t variable type found in
Linux's statfs(2), whose the man page states:
    The __fsword_t type used for various fields in the statfs structure
    definition is a glibc internal type, not intended for public use.  This
    leaves the programmer in a bit of a conundrum when trying to copy or
    compare these fields to local variables in a program.  Using unsigned int
    for such variables suffices on most systems.
    http://man7.org/linux/man-pages/man2/statfs.2.html
Due to this (as well as the MAGIC constants for various filesystems, such as
    #define SELINUX_MAGIC           0xf97cff8c
), we made our `struct ftyperec` use `unsigned int f_type` in de38471 on
2016-10-12.

Unfortunately, at least one Linux system (32-bit Debian wheezy from 2013, LTS
until 2018) uses a signed int for __fsword_t.  Explicitly casting `sfs.f_type`
to an `unsigned int` ensures that we're comparing two values with the same
signed-ness.

+++ lib/util/getfstype.c
@@ -222,7 +222,7 @@ getfstype(const char * path)
 	/* We need to call statfs(2) and interpret f_type values. */
 	if (statfs(path, &sfs) == 0) {
 		for (i = 0; ftypes[i].typename != NULL; i++) {
-			if (sfs.f_type == ftypes[i].f_type) {
+			if (((unsigned int)sfs.f_type) == ftypes[i].f_type) {
 				fstype = ftypes[i].typename;
 				break;
 			}
